Move components to individual folders
