+++ client/menu-item/style.scss
@@ -1,0 +1,3 @@
+.woocommerce-navigation__menu-item {
+	padding: $gap-smaller 0;
+}
