+++ client/navigation/index.js
@@ -6,7 +6,7 @@ import { Component } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import Menu from './Menu';
+import Menu from '../menu';
 
 export default class Navigation extends Component {
 	componentDidMount() {
