+++ client/navigation/style.scss
@@ -1,0 +1,24 @@
+.woocommerce-navigation {
+	color: initial;
+	position: fixed;
+	top: $admin-bar-height;
+	left: 0;
+	width: 272px;
+	height: 100%;
+	background: var(--wp-admin-sidebar-color);
+}
+
+.has-woocommerce-navigation {
+	#wpcontent, #wpfooter {
+		margin-left: 272px;
+	}
+
+	#adminmenuwrap,
+	#adminmenuback {
+		display: none !important;
+	}
+}
+
+.toplevel_page_woocommerce ul.wp-submenu {
+	display: none;
+}
